requirements.txt
# MCP Drug Discovery Pipeline - Python Requirements
# Generated from environment.yml for pip-only installations

# Core framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
typer>=0.9.0

# HTTP and async
httpx[http2]>=0.25.2
aiofiles>=23.2.0

# Data validation and configuration
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1

# Scientific computing
numpy>=1.24.4
pandas>=2.1.4
scipy>=1.11.0
scikit-learn>=1.3.2

# Network analysis
networkx>=3.2.1

# Molecular modeling
rdkit-pypi>=2023.9.2

# Bioinformatics
biopython>=1.82

# Visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
loguru>=0.7.2
tqdm>=4.66.1
requests>=2.31.0
openpyxl>=3.1.2

# Caching
redis>=5.0.1

# Development and testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.1
pre-commit>=3.6.0

# Jupyter for development
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Optional ML dependencies (uncomment if needed)
# torch>=2.0.0
# transformers>=4.35.0
# huggingface-hub>=0.17.0
# deepchem>=2.7.0
